Feat: #118 zustand 스토어에 slice pattern 적용
