+++ src/services/axiosProvider.ts
@@ -4,7 +4,7 @@ import { JWT_TOKEN_DUMMY } from '@mocks/mockData';
 import type { AxiosInstance, AxiosRequestConfig } from 'axios';
 import useToast from '@hooks/useToast';
 import { getAccessToken } from '@services/authService';
-import { useAuthStore } from '@stores/useAuthStore';
+import { useStore } from '@stores/useStore';
 
 const BASE_URL = import.meta.env.VITE_BASE_URL;
 const defaultConfigOptions: AxiosRequestConfig = {
@@ -29,7 +29,7 @@ export const authAxios = axiosProvider({
 // 요청 인터셉터
 authAxios.interceptors.request.use(
   (config) => {
-    const { accessToken } = useAuthStore.getState();
+    const { accessToken } = useStore.getState().auth;
 
     const modifiedConfig = { ...config };
     if (accessToken) modifiedConfig.headers.Authorization = `Bearer ${accessToken}`;
@@ -48,7 +48,7 @@ authAxios.interceptors.response.use(
     // 액세스 토큰 만료 시 처리
     if (error.response?.status === 401) {
       const { toastError } = useToast();
-      const { onLogout, setAccessToken } = useAuthStore.getState();
+      const { onLogout, setAccessToken } = useStore.getState().auth;
 
       // 에러 객체의 설정 객체 추출
       const originalRequest = error.config;
